+++ lib/storex/registries/ets.ex
@@ -11,7 +11,7 @@ defmodule Storex.Registry.ETS do
   @registry :storex_registry
 
   @doc false
-  def start_link do
+  def start_link(_) do
     GenServer.start_link(__MODULE__, nil, name: @registry)
   end